+++ pina/model/low_rank_neural_operator.py
@@ -1,29 +1,33 @@
-"""Module Averaging Neural Operator."""
+"""Module LowRank Neural Operator."""
 
 import torch
 from torch import nn, cat
-from .block import AVNOBlock
-from .base_no import KernelNeuralOperator
+
 from ..utils import check_consistency
 
+from .kernel_neural_operator import KernelNeuralOperator
+from .block.lowrank_block import LowRankBlock
 
-class AveragingNeuralOperator(KernelNeuralOperator):
+
+class LowRankNeuralOperator(KernelNeuralOperator):
     """
-    Implementation of Averaging Neural Operator.
+    Implementation of LowRank Neural Operator.
 
-    Averaging Neural Operator is a general architecture for
+    LowRank Neural Operator is a general architecture for
     learning Operators. Unlike traditional machine learning methods
-    AveragingNeuralOperator is designed to map entire functions
-    to other functions. It can be trained with Supervised learning strategies.
-    AveragingNeuralOperator does convolution by performing a field average.
+    LowRankNeuralOperator is designed to map entire functions
+    to other functions. It can be trained with Supervised or PINN based
+    learning strategies.
+    LowRankNeuralOperator does convolution by performing a low rank
+    approximation, see :class:`~pina.model.block.lowrank_layer.LowRankBlock`.
 
     .. seealso::
 
-        **Original reference**: Lanthaler S. Li, Z., Kovachki,
-        Stuart, A. (2020). *The Nonlocal Neural Operator:
-        Universal Approximation*.
-        DOI: `arXiv preprint arXiv:2304.13221.
-        <https://arxiv.org/abs/2304.13221>`_
+        **Original reference**: Kovachki, N., Li, Z., Liu, B.,
+        Azizzadenesheli, K., Bhattacharya, K., Stuart, A., & Anandkumar, A.
+        (2023). *Neural operator: Learning maps between function
+        spaces with applications to PDEs*. Journal of Machine Learning
+        Research, 24(89), 1-97.
     """
 
     def __init__(
@@ -32,8 +36,12 @@ class AveragingNeuralOperator(KernelNeuralOperator):
         projecting_net,
         field_indices,
         coordinates_indices,
-        n_layers=4,
-        func=nn.GELU,
+        n_kernel_layers,
+        rank,
+        inner_size=20,
+        n_layers=2,
+        func=torch.nn.Tanh,
+        bias=True,
     ):
         """
         :param torch.nn.Module lifting_net: The neural network for lifting
@@ -48,16 +56,26 @@ class AveragingNeuralOperator(KernelNeuralOperator):
             in the input tensor.
         :param list[str] coordinates_indices: the label of the
             coordinates in the input tensor.
-        :param int n_layers: number of hidden layers. Default is 4.
-        :param torch.nn.Module func: the activation function to use,
-            default to torch.nn.GELU.
+        :param int n_kernel_layers: number of hidden kernel layers.
+            Default is 4.
+        :param int inner_size: Number of neurons in the hidden layer(s) for the
+            basis function network. Default is 20.
+        :param int n_layers: Number of hidden layers. for the
+            basis function network. Default is 2.
+        :param func: The activation function to use for the
+            basis function network. If a single
+            :class:`torch.nn.Module` is passed, this is used as
+            activation function after any layers, except the last one.
+            If a list of Modules is passed,
+            they are used as activation functions at any layers, in order.
+        :param bool bias: If ``True`` the MLP will consider some bias for the
+            basis function network.
         """
 
         # check consistency
         check_consistency(field_indices, str)
         check_consistency(coordinates_indices, str)
-        check_consistency(n_layers, int)
-        check_consistency(func, nn.Module, subclass=True)
+        check_consistency(n_kernel_layers, int)
 
         # check hidden dimensions match
         input_lifting_net = next(lifting_net.parameters()).size()[-1]
@@ -77,7 +95,7 @@ class AveragingNeuralOperator(KernelNeuralOperator):
             != projecting_net_input
         ):
             raise ValueError(
-                "The projecting_net input must be equal to"
+                "The projecting_net input must be equal to "
                 "the embedding dimension (which is the output) "
                 "of the lifting_net plus the dimension of the "
                 "coordinates, i.e. len(coordinates_indices)."
@@ -87,15 +105,26 @@ class AveragingNeuralOperator(KernelNeuralOperator):
         self.coordinates_indices = coordinates_indices
         self.field_indices = field_indices
         integral_net = nn.Sequential(
-            *[AVNOBlock(output_lifting_net, func) for _ in range(n_layers)]
+            *[
+                LowRankBlock(
+                    input_dimensions=len(coordinates_indices),
+                    embedding_dimenion=output_lifting_net,
+                    rank=rank,
+                    inner_size=inner_size,
+                    n_layers=n_layers,
+                    func=func,
+                    bias=bias,
+                )
+                for _ in range(n_kernel_layers)
+            ]
         )
         super().__init__(lifting_net, integral_net, projecting_net)
 
     def forward(self, x):
         r"""
-        Forward computation for Averaging Neural Operator. It performs a
+        Forward computation for LowRank Neural Operator. It performs a
         lifting of the input by the ``lifting_net``. Then different layers
-        of Averaging Neural Operator Blocks are applied.
+        of LowRank Neural Operator Blocks are applied.
         Finally the output is projected to the final dimensionality
         by the ``projecting_net``.
 
@@ -108,11 +137,12 @@ class AveragingNeuralOperator(KernelNeuralOperator):
         :return: The output tensor obtained from Average Neural Operator.
         :rtype: torch.Tensor
         """
-        points_tmp = x.extract(self.coordinates_indices)
-        new_batch = x.extract(self.field_indices)
-        new_batch = cat((new_batch, points_tmp), dim=-1)
-        new_batch = self._lifting_operator(new_batch)
-        new_batch = self._integral_kernels(new_batch)
-        new_batch = cat((new_batch, points_tmp), dim=-1)
-        new_batch = self._projection_operator(new_batch)
-        return new_batch
+        # extract points
+        coords = x.extract(self.coordinates_indices)
+        # lifting
+        x = self._lifting_operator(x)
+        # kernel
+        for module in self._integral_kernels:
+            x = module(x, coords)
+        # projecting
+        return self._projection_operator(cat((x, coords), dim=-1))
