+++ pina/problem/__init__.py
@@ -8,6 +8,6 @@ __all__ = [
 
 from .abstract_problem import AbstractProblem
 from .spatial_problem import SpatialProblem
-from .timedep_problem import TimeDependentProblem
+from .time_dependent_problem import TimeDependentProblem
 from .parametric_problem import ParametricProblem
 from .inverse_problem import InverseProblem
