+++ pina/solver/reduced_order_model.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from pina.solver import SupervisedSolver
+from . import SupervisedSolver
 
 
 class ReducedOrderModelSolver(SupervisedSolver):
