+++ tests/test_blocks/test_spectral_convolution.py
